requirements.txt
fastapi[all]==0.87.0
scikit-learn==1.1.3
pandas==1.5.1
mlflow==2.0.1
sqlalchemy==1.4.44
psycopg2-binary==2.9.5
numpy==1.23.5
